chore: merge new and existing utils files
