+++ src/js/analyses/components/Charts/CoverageChart.tsx
@@ -6,7 +6,7 @@ import { area } from "d3-shape";
 import React, { useEffect, useRef } from "react";
 import styled, { DefaultTheme } from "styled-components";
 import { theme } from "../../../app/theme";
-import { deriveTrustworthyRegions } from "./utils";
+import { deriveTrustworthyRegions } from "../../utils";
 
 function draw(element, data, length, yMax, untrustworthyRanges) {
     select(element).append("svg");
