+++ src/js/analyses/components/Iimi/CondensedIimiCoverage.tsx
@@ -1,21 +1,24 @@
-import { filter, map, unzip } from "lodash-es";
+import { filter, map, sortBy, unzip } from "lodash-es";
 import React from "react";
 import { IimiIsolate, IimiSequence } from "../../types";
+import { averageSequences, combineUntrustworthyRegions, convertRleToCoverage } from "../../utils";
 import { SummaryChart } from "../Charts/SummaryChart";
-import { averageSequences, combineUntrustworthyRegions, convertRleToCoverage } from "../Charts/utils";
 
 /** A graph of averaged Iimi analysis sequence coverage */
 export function CondensedIimiCoverage({ isolates }: { isolates: IimiIsolate[] }) {
-    const sequences = unzip(map(isolates, "sequences"));
+    const sequences = sortBy(unzip(map(isolates, "sequences")), seqs => seqs[0]?.length);
 
     const charts = map(sequences, (seqs: IimiSequence[]) => {
+        const filteredSeqs = filter(seqs);
         const avgSeq = averageSequences(
-            map(filter(seqs), (seq: IimiSequence) => {
+            map(filteredSeqs, (seq: IimiSequence) => {
                 return convertRleToCoverage(seq.coverage.lengths, seq.coverage.values);
             }),
         );
+
         return (
             <SummaryChart
+                key={filteredSeqs[0].id}
                 data={avgSeq}
                 id={isolates[0].id}
                 yMax={Math.max(...avgSeq, 10)}
